+++ packages/base-driver/lib/express/middleware.js
@@ -57,13 +57,11 @@ export function allowCrossDomainAsyncExecute(basePath) {
  * @returns {any}
  */
 export function handleLogContext(req, res, next) {
-  const requestId = util.uuidV4();
+  const requestId = fetchHeaderValue(req, 'x-request-id') || util.uuidV4();
 
   const sessionId = SESSION_ID_PATTERN.exec(req.url)?.[1];
   const sessionInfo = sessionId ? {sessionId, sessionSignature: calcSignature(sessionId)} : {};
-  const isSensitiveHeaderValue = _.isArray(req.headers['x-appium-is-sensitive'])
-    ? _.first(req.headers['x-appium-is-sensitive'])
-    : req.headers['x-appium-is-sensitive'];
+  const isSensitiveHeaderValue = fetchHeaderValue(req, 'x-appium-is-sensitive');
 
   log.updateAsyncContext({
     requestId,
@@ -140,4 +138,15 @@ export function catch404Handler(req, res) {
   log.debug(`No route found for ${req.url}`);
   const [status, body] = getResponseForW3CError(new errors.UnknownCommandError());
   res.status(status).json(body);
+}
+
+/**
+ * @param {import('express').Request} req
+ * @param {string} name
+ * @returns {string | undefined}
+ */
+function fetchHeaderValue(req, name) {
+  return _.isArray(req.headers[name])
+    ? req.headers[name][0]
+    : req.headers[name];
 }